+++ src/commands/shared-links/create.js
@@ -10,7 +10,7 @@ class SharedLinksCreateCommand extends BoxCommand {
 
 		let sharedLinksModule = new SharedLinksModule(this.client);
 		let updatedItem = await sharedLinksModule.createSharedLink(args, flags);
-		await this.output(updatedItem.shared_link);
+		await this.output(updatedItem);
 	}
 }